moved routes to seperate file
