+++ TODO.txt
@@ -1,3 +1,3 @@
 TODO:
-Seperate index.js to multiple files
+Seperate main.js to multiple files
 Add apks again
