+++ start_server.bash
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+node -r dotenv/config main.js
